Add Users to database
Created and ran migration to add Users table to database. For the sake of consistency in attempting to reproduce the bug, the Users table will be part of a single table inheritance structure: Users will be one of two types.

Ran 'bin/rails g migration CreateUsers' and added column information to the migration.
Then ran 'bin/rails db:migrate' to update the schema.